fixed file order: added index.ts files, fixed imports and ets. minor changes in styles
added index.ts for constants

file order is fixed. minor style changes
