+++ src/common/Button/Button.tsx
@@ -1,7 +1,4 @@
-interface ButtonProps {
-  buttonText: string,
-  handleClick: () => void
-}
+import { ButtonProps } from "./types"
 
 export const Button = ({buttonText, handleClick}: ButtonProps) => {
   return <button onClick={handleClick}>{buttonText}</button>
