+++ src/common/Button/index.ts
@@ -1,0 +1,2 @@
+export { Button } from './Button'
+export { Input } from './Input'
