+++ src/common/Button/types.ts
@@ -1,0 +1,11 @@
+export interface ButtonProps {
+  buttonText: string,
+  handleClick: () => void
+}
+
+export interface InputProps {
+  value: string,
+  placeholderText: string,
+  labelText?: string,
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void
+}
