+++ src/common/SearchBar/styled.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components"
+
+export const SearchPanelWrapper = styled.div`
+  display: flex;
+  justify-content: start;
+  width: 1122px;
+  margin-bottom: 20px;
+`
