+++ src/common/SearchBar/types.ts
@@ -1,0 +1,5 @@
+export interface SearchBarProps {
+  handleSearch: (searchInputValue: string) => void,
+  searchValue: string,
+  handleSearchInput: (e: React.ChangeEvent<HTMLInputElement>) => void,
+}
