+++ src/components/CourseInfo/CourseInfo.tsx
@@ -1,36 +1,9 @@
-import styled from "styled-components";
-import { Button } from "../../common/Button/Button";
+import { Button } from "../../common/Button";
 import { getAuthorNames, formatDuration } from "../../helpers";
-import { Course } from "../Courses/Courses";
-import { strings } from "../../strings";
-
-// type AuthorList = {
-//   id: string,
-//   name: string
-// }
-
-interface CourseInfoProps {
-  coursesList: Course[],
-  // authorsList: AuthorList[],
-  onBack: () => void,
-  showCourseId: string,
-}
-
-const CourseInfoWrapper = styled.div`
-  display: flex;
-  width: 1122px;
-  height: 240px;
-  padding: 20px;
-  border: 2px solid #007298;
-  border-left: 8px solid #007298;
-  border-radius: 10px;
-  margin-bottom: 10px;
-  box-shadow: 4px 4px 10px #007298;
-  justify-content: space-between;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 20px;
-`
+import { CourseInfoProps } from "./types";
+import { dictionary } from "../../strings";
+import { CourseInfoWrapper, InfoSectionStyled, DescriptionStyled } from './styled';
+import { mockedAuthorsList } from "../../constants/constants";
 
 export const CourseInfo = ({ coursesList, onBack, showCourseId }: CourseInfoProps) => {
   const getCourse = (id: string) => {
@@ -40,18 +13,18 @@ export const CourseInfo = ({ coursesList, onBack, showCourseId }: CourseInfoProp
   return (
     <>
       <CourseInfoWrapper>
-        <section>
+        <DescriptionStyled>
           <h2>{getCourse(showCourseId)?.title}</h2>
           <p>{getCourse(showCourseId)?.description}</p>
-        </section>
-        <section>
+        </DescriptionStyled>
+        <InfoSectionStyled>
           <p><b>ID: </b>{getCourse(showCourseId)?.id}</p>
           <p><b>Duration: </b>{formatDuration(getCourse(showCourseId)?.duration)}</p>
           <p><b>Created: </b>{getCourse(showCourseId)?.creationDate}</p>
-          <p><b>Authors: </b>{getAuthorNames(getCourse(showCourseId)?.authors)}</p>
-        </section>
+          <p><b>Authors: </b>{getAuthorNames(getCourse(showCourseId)?.authors, mockedAuthorsList)}</p>
+        </InfoSectionStyled>
       </CourseInfoWrapper>
-      <Button buttonText={strings.buttonBack} handleClick={onBack} />
+      <Button buttonText={dictionary.buttonBack} handleClick={onBack} />
     </>
 
   )
