+++ src/components/CourseInfo/index.ts
@@ -1,0 +1,1 @@
+export {CourseInfo} from './CourseInfo'
