+++ src/components/CourseInfo/styled.ts
@@ -1,0 +1,29 @@
+import styled from "styled-components"
+
+export const CourseInfoWrapper = styled.div`
+  display: flex;
+  width: 1122px;
+  height: 240px;
+  padding: 20px;
+  border: 2px solid #007298;
+  border-left: 8px solid #007298;
+  border-radius: 10px;
+  margin-bottom: 10px;
+  box-shadow: 4px 4px 10px #007298;
+  justify-content: space-between;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 20px;
+`
+
+export const DescriptionStyled = styled.section`
+  width: 677px;
+  border-right: solid 1px;
+  padding: 10px;
+`
+
+export const InfoSectionStyled = styled.section`
+  width: 323px;
+  display: flex;
+  flex-direction: column;
+`
