+++ src/components/CourseInfo/types.ts
@@ -1,0 +1,7 @@
+import { CourseType } from "../Courses/types";
+
+export interface CourseInfoProps {
+  coursesList: CourseType[],
+  onBack: () => void,
+  showCourseId: string,
+}
