+++ src/components/Courses/CourseCard/index.ts
@@ -1,0 +1,1 @@
+export { CourseCard } from './CourseCard'
