+++ src/components/Courses/CourseCard/types.ts
