+++ src/components/Courses/index.ts
@@ -1,0 +1,2 @@
+export { Courses } from './Courses'
+export { EmptyCourseList } from './EmptyCourseList'
