+++ src/components/Courses/styled.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components"
+
+export const EmptyListWrapperStyled = styled.div`
+  height: 100vh;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`
