+++ src/components/Header/Header.tsx
@@ -1,27 +1,9 @@
-import { Button } from "../../common/Button/Button";
-import { Logo } from "./Logo";
-import styled from "styled-components";
+import { Button } from "../../common/Button";
+import { Logo } from "../Logo";
+import { HeaderProps } from "./types";
+import { HeaderStyled, LoginWrapper } from "./styled";
 
-const HeaderStyled = styled.header`
-position: sticky;
-width: 100%;
-top: 0;
-left: 0;
-height: 65px;
-display: flex;
-justify-content: space-between;
-background-color: #fff;
-padding-left: 10px;
-`;
-
-const LoginWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  position: relative;
-  right: 40px;
-`
-
-export const Header = ({ handleClick }) => {
+export const Header = ({ handleClick }: HeaderProps) => {
   return (
     <HeaderStyled>
       <Logo />
