+++ src/components/Header/styled.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const HeaderStyled = styled.header`
+position: sticky;
+width: 100%;
+top: 0;
+left: 0;
+height: 65px;
+display: flex;
+justify-content: space-between;
+background-color: #fff;
+padding-left: 10px;
+`
+
+export const LoginWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  position: relative;
+  right: 40px;
+`
