+++ src/components/Header/types.ts
@@ -1,0 +1,3 @@
+export type HeaderProps = {
+  handleClick: () => void
+}
