+++ src/components/Logo/Logo.tsx
@@ -1,0 +1,10 @@
+import logo from '../../../src/assets/logo.png';
+import { LogoImageStyled } from './styled';
+
+export const Logo = () => {
+  return (
+    <div>
+      <LogoImageStyled src={logo} alt="logo" />
+    </div>
+  )
+}
