+++ src/components/Logo/styled.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components"
+
+export const LogoImageStyled = styled.img`
+height: 100%;
+width: auto;
+`
