+++ src/helpers/types.ts
@@ -1,0 +1,4 @@
+export type authorsListType = {
+  id: Readonly<string>,
+  name: Readonly<string>,
+}
